Make puzzle mode work, fix bugs
